+++ src/drapo/main.py
@@ -4,8 +4,7 @@ import logging
 import sys
 import time
 import schedule
-from drapo.utils.cli_parser import parser
-from drapo.utils.toml_parser import load_orchestration_config
+from drapo.utils import *
 from drapo.orchestrer import run_flow, schedule_jobs
